delete unused test and moe threshold bitmap
