+++ app/src/main/java/com/scepticalphysiologist/dmaple/map/creator/MapCreator.kt
@@ -4,7 +4,7 @@ import android.graphics.Bitmap
 import android.graphics.Rect
 import android.util.Size
 import com.scepticalphysiologist.dmaple.etc.Edge
-import com.scepticalphysiologist.dmaple.etc.ThresholdBitmap
+import com.scepticalphysiologist.dmaple.ui.camera.ThresholdBitmap
 import com.scepticalphysiologist.dmaple.map.field.FieldRoi
 import mil.nga.tiff.FileDirectory
 import java.lang.IllegalArgumentException
