--- app/src/test/java/com/scepticalphysiologist/dmaple/map/creator/MapCreatorTest.kt
@@ -1,23 +1,0 @@
-package com.scepticalphysiologist.dmaple.map.creator
-
-import android.graphics.Bitmap
-import org.junit.jupiter.api.Test
-import org.junit.runner.RunWith
-import org.robolectric.RobolectricTestRunner
-
-@RunWith(RobolectricTestRunner::class)
-class MapCreatorTest {
-
-    @Test
-    fun `seed spine`() {
-
-        val bitmap = Bitmap.createBitmap(400, 200, Bitmap.Config.ARGB_8888)
-
-
-    }
-
-
-
-
-
-}
